+++ CHANGELOG.md
@@ -1,6 +1,14 @@
+## v.2.0 (May 15, 2016)
+
+[jekyll-pandoc 2.0](https://github.com/mfenner/jekyll-pandoc/releases/tag/v.2.0) was released on September 5, 2015 with the following changes:
+
+* require Jekyll 3.0 or higher
+* use `pandoc-ruby` 2.0
+* removed jekyll monkey patching
+
 ## v.1.0 (September 5, 2015)
 
-[jekyll-pandoc 1.0](https://github.com/datacite/omniauth-orcid/releases/tag/v.1.0) was released on September 5, 2015 with the following changes:
+[jekyll-pandoc 1.0.2](https://github.com/mfenner/jekyll-pandoc/releases/tag/v.1.0.2) was released on September 5, 2015 with the following changes:
 
 * use Jekyll 2 compatible errors
 * require pandoc-ruby gem `1.0.0` or higher